Initial project setup and schema definition
Set up a vite/react/typescript project with tailwind for the
frontend, and a python backend with FastAPI connected to a
postgres db. User docker for the backend with a docker-compose
to run the backend and db at the same time.

Define some db models for an item, deal, recipe and ingredient.